Added a navigation drawer to the main layout.
The navigation drawer is the primary way to navigate through the various screens in Hubble.
The content is wrapped in the `main` element. The left and top margins of the `main`
element are adjusted such that the drawer and toolbar do not overlap the content. Currently,
the margins are hardcoded. This should be updated to be dynamic.