"New Account" dialog invokes `createAccount` action creator on save.
The Redux Thunk and Axios packages were added as dependencies. The `createAccount()` action creator
makes a REST API call via Axios to create an account on the server. It dispatches a notification at
the beginning and the end of the request. Therefore, the required action creators and reducers for
showing notifications were added.

The `RecordForm` component was updated to invoke a listener provided by the parent component,
mostly a controller, to notify whenever the fields are modified by the user. Further, it now
requires a `values` prop which contains the current values of the fields. The parent component
is responsible for extracting these values from the `groups` props that describes the layout
of the form. A utility function `extractValues()` was added for this purpose.