Created the "View Accounts" page.
The `ViewAccounts` component loads the accounts by dispatching the asynchronous `fetchAccounts()`
action. The `showNotification()` and `fetchAccountsComplete()` actions are dispatched at the end
and beginning, respectively, after the data is fetched. The corresponding reducers were added
to update the shape of the store.

The `WorkspaceTable`, `WorkspaceTableToolbar` and `WorkspaceTableHead` occupy the majority of the
`ViewAccounts` page. These components are generic enough such that other workspace pages can make
use of them.