Implemented subscription and transaction dialogs in the quick add menu
Dialogs for NewSubscription and NewTransaction are added to the quick add menu. Reducers for NEW_SUBSCRIPTION, NEW_TRANSACTION as well as NEW_PLAN and NEW_INVOICE actions have been added in this commit to prevent a minor bug (that caused NewTransaction dialog to not be displayed) in future commits of adding new dialogs.