Added a fake subscription generator in the mock backend.
Two utility functions, `soonDate()` and `pastDate()`, were added to create random ISO date strings.
The documentation of Faker.js specifies that `faker.date.soon()` is exported. However, on inspecting
the module, it appears that `soon()` is not available. I think we are using an older version of
Faker.js. We may want to update this in the future. Right now, `soonDate()` returns a future date.