Linked View Accounts and View Account pages.
When the user click on a row in View Accounts page the View Account page is opened with the identifier
of the account. View Account page then makes a REST API call to fetch the account details and render
it. This required introduction of an asynchronous action and a reducer.

The behavior of the WorkspaceTable component had be changed to accomodate this. The row checkbox
uses a different handler than the other cells. Right now, each cell receives a event handler.
There was an issue with using just two handlers: one for checkbox and another one for the whole
row. However, this did not work because of some event propogation logic that I am not able to
figure out.

A bug was found and fixed in the mock backend. Basically, the first capture returned by `Regex.exec()`,
is the regex itself. Previously, the first capture was being used as the identifier parameter.