Clear account from the store when the screen changes.
The View Account page renders a circular progress when the account in the store is null. However,
if the store has an account from a previous visit to the page, it temporarily renders the stale
account. This may confuse the user. Therefore, the account is cleared from the store when the
user leaves the page.