AccountFormDialog is injected with the values of the account being edited.
The state of the AccountFormDialog component was significantly modified in this commit.
Previously, a two dimensional array was used to keep track of all the values. However,
now we use key-value pairs. RecordForm was modified to keep up with these changes.
In doing so, the functionality of the other add dialogs were obstructed.

The `EditAccount` component is a simple container that renders AccountFormDialog with
appropriate callbacks. We could have done the same within the bounds of `MainLayout`.
However, that would have meant for the layout to be tightly coupled with all the
dialogs it renders. Therefore, the functionality was abstracted into `EditAccount`.
Similar abstractions will be created shortly.