Updated `SubscriptionFormDialog` to be more flexible.
Since the `NewSubscription` component could be reused for editing subscriptions, it was renamed to
`SubscriptionFormDialog`.

The internal state of this component was significantly modified in this commit. Previously a two
dimensional array was used to keep track of all the values. Now we use key-value pairs. This change
was made to keep up with the `RecordForm`. However, the shape of the form state is somewhat
transparent because we use `extractValues` exported by the `RecordForm.js` module, which takes
care of extracting the default state based on the form configuration.