Implemented the View Subscriptions page.
This page was derived from the View Accounts page by making appropriate modifications.
The rows in the `WorkspaceTable` component were rendered specifically for accounts.
In order to make to generic, the `WorkspaceTable` was updated to accept the `renderCellValue`
prop, a callback function responsible for rendering the cell. This modification has caused the
View Accounts page to crash the application. This issue will be fixed shortly.