Implemented the View Plans page.
This page was derived from the View Accounts page by making appropriate modifications.
The `fetchPlansComplete()` and `fetchPlans()` action creators were introduced to fetch
the plans from the backend. The `plansReducer()` function updates the store with
the plans fetched from the backend.

The `WorkspaceFilter` component was removed from the View Plans page. Further, the
`WorkspaceTable` now accepts a `rowsPerPage` prop.