Created the View Plan page.
The View Plans page is now linked to View Plan page. When the user clicks on a plan in View Plans,
the user is redirected to the View Plan page, using the identifier of the plan in question to create
the URL. The identifier is then used to fetch the plan from the backend using `fetchPlan()`
asynchronous action creator.