The `Edit` button in the `PlanCard` component now opens `PlanFormDrawer`.
The `editPlan()` action creator dispatches the `EDIT_PLAN` action with the
plan as payload. The `openDialogReducer` prepares the store for rendering the
dialog and `planReducer` stores the plan to edit.

When the user clicks on the Save button, edits are sent to the backend for
persistence. The `savePlan()` asynchronous action creator dispatches the
`FETCH_PLAN_COMPLETE` action on receving response from the backend. Thus,
the fields in View Plan page is automatically updated.