Created the View Invoice page.
The View Invoices and View Invoice pages are now linked. When the user clicks on an invoice in the
View Invoices page they are redirected to the View Invoice page, using the identifier of the invoice
in question to create the URL. The identifier is then used to fetch the invoice from the backend using
`fetchInvoice()` asynchronous action creator. The fetched invoice is then passed to the`InvoiceCard`
component for rendering.