The edit button in the View Invoice page now opens the InvoiceFormDrawer.
The `editInvoice()` action creator dispatches the `EDIT_INVOICE` action with the
invoice as payload. The `openDialogReducer` prepares the store for rendering the
dialog and `inviceReducer` stores the invoice to edit.

When the user clicks on the Save button, edits are sent to the backend for
persistence. The `saveInvoice()` asynchronous action creator dispatches the
`FETCH_INVOICE_COMPLETE` action on receving response from the backend. Thus,
the fields in View Invoice page is automatically updated.