The View Account page now generates URL based on the workspace filter.
Two utility functions `toURLParams()` and `toFilterState()` were added to implement this.
The former function converts the workspace filter state to an object that can be converted
to URL param using the `URLSearchParams` class. On the other hand latter function substitutes
`extractFilterState()`. It merges both the default values from the filter configuration and
the URL params provided as the second argument. The URL params always override the default
values. This will allow us to initialize `ViewAccounts` component based on the URL.

The clear button not only defaults the workspace filter but also changes the URL of ViewAccounts
page to `/accounts` by removing all the filter parameters. This behavior will be implemented
in the other pages soon.

Finally, the dates sent from the mock backend are slowly being migrated to use the ISO format.
To do this we simply invoke the `toISOString() ` on the `Date` object. A utility function
`toDateString()` in `utils.js` helps us convert dates to locale specific dates. As of this
writing, only the View Accounts page is aware of this.