Added the View Transaction page.
When the user clicks on a transaction in the View Transactions page they are redirected to the View Transaction page, using the identifier of the transaction in question to create the URL. The identifier is then used to fetch the transaction from the backend using fetchTransaction() asynchronous action creator. The fetched transaction is then passed to the TransactionCard component for rendering.