The View Accounts page updates the values of the workspace filter based on the URL parameters.
The utility functions `toURLParams()` and `toFilterState()` have been moved to `WorkspaceFilter.js`.
`extractFilterState()` will be shortly removed.

The query-string dependency was added to the project. It is used to extract the URL parameters
packed in an object. The object is then fed to `toFilterState()` which merges the URL parameters
and the default values specified in the form configuration, with the former having more priority.
The workspace filter is shown by default if the URL has any parameters, regardless of being related
to the filter or not. Similarly, it is hidden when there are no parameters. A weird behavior is
manifested when the clear button is clicked in the filter. It does not hide the workspace filter
even though it causes the URL to lose all the parameters. I think this has to do something with
React hooks, in particular the `useState()` hook. Luckily for us, this is the behavior we wanted
to implement.