Created the Edit Transaction dialog.
The edit button in the View Transaction page now opens the TransactionFormDrawer. The editTransaction() action creator dispatches the EDIT_TRANSACTION action with the transaction as payload. The openDialogReducer prepares the store for rendering the dialog and transactionReducer stores the transaction to edit. When the user clicks on the Save button, edits are sent to the backend for persistence. The saveTransaction() asynchronous action creator dispatches the FETCH_TRANSACTION_COMPLETE action on receving response from the backend. Thus, the fields in View Transaction page is automatically updated.