View Accounts page now requests the backend based on the filter values.
It should be noted that the mock backend does not acknowledge the filter parameters.
However, it was verified that the filters parameters reached the mock backend,
which means when the real backend is integrated the filters will work as expected.

Also, removed the "tatuq" file. We could time travel to find its origin. But nah.