Added the Net Subscriber Change card.
The `nivo` package has certain properties that have to be explicitly stated, otherwise they aren't rendered. For example, the height and width only stipulate the corresponding property of the plot itself, and the legends and axis points are usually out of this height x width area, which can only be rendered by stating the margin separately. This might be useful while adding other components from `nivo`.