Added the Analytics page
The Analytics page now has 9 components. The components are completely finished yet. The simplest of these components are cards that display the total number of a type of object, such as `AccountsTotalCard` or `TransactionsTotalCard`. There are 5 of these cards that display a single number. The remaining four components are graphs from the `nivo` package. Only three types of graphs were used to make this initial version of the Analytics page - Line graph, Bar Graph and Pie Chart.