Added unit tests for all the action creators.
I should have probably divided this commit into multiple smaller commits. Perhaps on the basis of
the domain the tests were written for.

A general pattern was used to create the tests for the action creators. As of this writing, each
domain has nine action creators (in most cases at least). Out of which, five are synchronous
action creators. Since they are pure functions, the tests are straightforward. As for the remaining
asynchronous the `axios-mock-adapter` and `redux-mock-store` packages are used. The basic idea is the
asynchronous action creator is dispatched against a mock adapter. On resolution of the promise
returned by the asynchronous action creator, we test the actions dispatched and the history using
the backend and store mocks.

For the most part, the tests were derived by making appropriate changes to the original tests that
I wrote for the accounts domain.

Further, a few missing action creators were added. Right now, these action creators are not being
used. However, in the future we will requrie them. With that in mind, I implemented them and wrote
their respestive tests here.