Removed the company group from the account form.
Further, a small bug in the View Accounts, View Subscriptions, View Invoices, and View Transactions
was fixed. The `Grid` that contains the workspace filter and workspace table was assigned to seperate
the chidren with a spacing of 2. This caused the horizontal scrollbar to appear. The spacing was
removed. The workspace filter now has a left margin of 16.