Added the Preferences page.
Clicking on a tab changes the `tabIndex`, a prop that is given to the tab panel. Each panel is an indexed page. Therefore, adding new tabs later on is very easy. Just index them in the main Preferences page, and make a new page for the tab that has to be added.