Added revenue and plans sections to analytics page.
In the revenue section, we have two graphs. The first one is a general graph about how much revenue is flowing in, how much is billed and how much is due. The second graph is a breakdown of how the revenue is collected - online, credit card, debit card, cash, etc.

The plans section has a plans break down graph which shows the top plans and their contributions to the whole. The second is a conversions breakdown - how many plans were converted in the last month, how many were cancelled, how many are not cancelled but yet to pay, recurring conversions, one time conversions.