Updated the forms in the Preferences page.
The preferences module was moved to the workspace directory. The containing directory was deleted.
At this point, it makes sense to group all the pages under the workspace directory. We will soon
move the analytics page here as well.

The content of each tab in the Preferences page is rendered using the `RecordForm` component.
The state of the form is managed by a new component called `PreferenceForms`, which wraps
`RecordForm` under the hood. Right now, groups refer to pages in RecordForm lingo. However,
`RecordForm` will be updated to differentiate between groups and pages soon. Further, a new
field type `password` was introduced to `RecordForm` in this commit.